flask-app/requirements.txt
boto3==1.24.4
botocore==1.27.4
jmespath==1.0.0
python-dateutil==2.8.2
python-dotenv==0.20.0
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.9
